a bunch of shit im not documenting

+++ mergesort.h
@@ -1,6 +1,6 @@
 #ifndef MERGESORT_H
 #define MERGESORT_H
 
-void mergesort(int sizeBig, int bigArray[], bool ascending, int &counter);
+void mergesort(int sizeBig, int bigArray[], bool ascending, long long &counter);
 
 #endif
